Create flames.py
Flames game is a method to find out the status of a love relationship. Entering two names will give you the outcome of a relationship between them.

FLAMES is a Love Test, a good old game of love that has helped lovers determine the scope of their love and relationship.